requirements.txt
django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.4.0
drf-spectacular==0.28.0
